+++ database/init.sql
@@ -3,14 +3,14 @@ CREATE TABLE IF NOT EXISTS containers (
     ip VARCHAR(31) NOT NULL,
     status VARCHAR(31) NOT NULL,
     last_ping_time TIMESTAMP NOT NULL,
-    ping_duration INTERVAL NOT NULL
+    ping_duration BIGINT
 );
 
-CREATE TABLE IF NOT EXISTS container_ping_history (
-    id SERIAL PRIMARY KEY,
-    container_id INTEGER REFERENCES containers(id) ON DELETE CASCADE,
-    ip VARCHAR(31) NOT NULL,
-    status VARCHAR(31) NOT NULL,
-    ping_time TIMESTAMP NOT NULL,
-    ping_duration INTERVAL NOT NULL
-)
+-- CREATE TABLE IF NOT EXISTS container_ping_history (
+--     id SERIAL PRIMARY KEY,
+--     container_id INTEGER REFERENCES containers(id) ON DELETE CASCADE,
+--     ip VARCHAR(31) NOT NULL,
+--     status VARCHAR(31) NOT NULL,
+--     ping_time TIMESTAMP NOT NULL,
+--     ping_duration BIGINT
+-- )