+++ frontend/docker-monitoring-frontend/src/api/api.ts
@@ -1,11 +1,13 @@
 import axios from 'axios';
 
-// TODO: замени на 'http://backend:8080/containers' когда поместишь frontend в docker контейнер
-// const API_URL = 'http://backend:8080/containers';
-// Для теста на локальном компьютере:
-const API_URL = 'http://localhost:8080/containers';
+const API_URL = process.env.REACT_APP_BACKEND_URL || 'http://localhost:8080';
+const BACKEND_URL = API_URL + "/containers"
 
 export const getPingResults = async () => {
-    const response = await axios.get(API_URL);
+    // if (!API_URL) {
+    //     throw new Error("Backend URL is not defined in the environment variables.");
+    // }
+    console.log("FULL_ROUTE_URL = ", BACKEND_URL)
+    const response = await axios.get(BACKEND_URL);
     return response.data;
 };