+++ frontend/docker-monitoring-frontend/src/models/ContainerPingResult.ts
@@ -2,6 +2,6 @@ export interface ContainerPingResult {
     id: number;
     ip: string;
     status: string;
-    lastPingTime: string;
-    pingDuration: number;
+    last_ping_time: string;
+    ping_duration: number;
 }