+++ CHANGELOG.md
@@ -56,3 +56,40 @@
 ### Изменено
 - Использованы параметры из enviroment (описано в docker-compose) контейнеров
   вместо захардкоженных переменных
+
+## [0.9.2] - 08-02-2025
+### Изменено
+- Теперь frontend отображает last ping time и ping duration корректно
+
+## [0.10.0] - 08-02-2025 - ветка containerize-frontend
+### Добавлено
+- ветка containerize-frontend
+- контейнеризация frontend (пока frontend не видит backend)
+
+## [0.10.1] - 08-02-2025 - ветка containerize-frontend
+### Изменено
+- убран volumes для сервиса db из docker-compose.yml
+- **Запускай с очищением кэша и томов volume** (чтобы избежать ошибки с init.sql):
+  ```bash
+  docker-compose down --volumes
+  docker-compose up --build
+  ```
+
+## [0.10.2] - 08-02-2025 - ветка containerize-frontend
+### Изменено
+- добавлен yarn.lock в репозиторий (и убран из .gitignore)
+
+## [0.10.3] - 08-02-2025 - ветка containerize-frontend
+### Изменено
+- разрешен конфликт с политикой CORS как для frontend (поднимаемом в контейнере), так и на локальном frontend сервисе
+
+## [0.10.4] - 08-02-2025 - ветка containerize-frontend -> (merge) main
+### Изменено
+- в postgres ping_duration теперь храним в наносекундах как BIGINT
+- во frontend отображаем в ms (делим на 1e6)
+- container_handler отправляет просто в наносекундах (ns)
+
+**merge containerize-frontend в main!**
+
+
+